refactor: move filter schemas to their own files
