+++ packages/x-adapter-platform/src/schemas/filters/number-filter.schema.ts
@@ -1,0 +1,17 @@
+import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
+import { NumberRangeFilter } from '@empathyco/x-types';
+import { PlatformFacetFilter } from '../../types';
+
+export const numberFilterSchema: Schema<PlatformFacetFilter, NumberRangeFilter> = {
+  id: 'filter',
+  facetId: 'id',
+  label: 'value',
+  selected: () => false,
+  modelName: () => 'NumberRangeFilter',
+  range: {
+    min: ({ value }) => Number(value.split('-')[0]),
+    max: ({ value }) => Number(value.split('-')[1])
+  }
+};
+
+export const numberFilterMutableSchema = createMutableSchema(numberFilterSchema);
